# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Staking Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/me/onboarding:
    post:
      operationId: completeOnboarding
      tags: [users]
      summary: Complete onboarding with full profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingInput"
      responses:
        "200":
          description: Onboarding complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, clerkId, email, role, availableBalance, totalEarnings, referralRewards, referralCode, onboardingComplete, isLocked, createdAt]
      properties:
        id:             { type: integer }
        clerkId:        { type: string }
        email:          { type: string }
        fullName:       { type: ["string", "null"] }
        location:       { type: ["string", "null"] }
        mpesaNumber:    { type: ["string", "null"] }
        role:           { type: string, enum: [USER, ADMIN] }
        availableBalance:  { type: number }
        totalEarnings:     { type: number }
        referralRewards:   { type: number }
        referralCode:      { type: string }
        onboardingComplete: { type: boolean }
        isLocked:          { type: boolean }
        createdAt:         { type: string, format: date-time }
    UserProfileUpdate:
      type: object
      properties:
        fullName:    { type: string }
        location:    { type: string }
        mpesaNumber: { type: string }
    OnboardingInput:
      type: object
      required: [fullName, location, mpesaNumber]
      properties:
        fullName:    { type: string }
        location:    { type: string }
        mpesaNumber: { type: string }
...