# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Staking Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: number;
  clerkId: string;
  email: string;
  /** @nullable */
  fullName?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  mpesaNumber?: string | null;
  role: UserRole;
  availableBalance: number;
  totalEarnings: number;
  referralRewards: number;
  referralCode: string;
  onboardingComplete: boolean;
  isLocked: boolean;
  createdAt: Date;
}