# lib/db/src/schema/users.ts
import { pgTable, text, serial, timestamp, boolean, numeric, pgEnum } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const roleEnum = pgEnum("role", ["USER", "ADMIN"]);

export const usersTable = pgTable("users", {
  id: serial("id").primaryKey(),
  clerkId: text("clerk_id").notNull().unique(),
  email: text("email").notNull().unique(),
  fullName: text("full_name"),
  location: text("location"),
  mpesaNumber: text("mpesa_number"),
  role: roleEnum("role").notNull().default("USER"),
  availableBalance: numeric("available_balance", { precision: 15, scale: 2 }).notNull().default("0"),
  totalEarnings: numeric("total_earnings", { precision: 15, scale: 2 }).notNull().default("0"),
  referralRewards: numeric("referral_rewards", { precision: 15, scale: 2 }).notNull().default("0"),
  referralCode: text("referral_code").notNull().unique(),
  referredBy: text("referred_by"),
  onboardingComplete: boolean("onboarding_complete").notNull().default(false),
  isLocked: boolean("is_locked").notNull().default(false),
  createdAt: timestamp("created_at", { withTimezone: true }).notNull().defaultNow(),
  updatedAt: timestamp("updated_at", { withTimezone: true }).notNull().defaultNow().$onUpdate(() => new Date()),
});

export const insertUserSchema = createInsertSchema(usersTable).omit({ id: true, createdAt: true, updatedAt: true });
export type InsertUser = z.infer<typeof insertUserSchema>;
export type User = typeof usersTable.$inferSelect;